Changed names around for Privacy and Terms, added content for Terms of Service
