+++ components/TermsOfService/Index.jsx
@@ -1,60 +1,51 @@
 import React from 'react';
 
-function PrivacyNotice() {
+function TermsOfService() {
   return (
     <div>
-      <h2>Privacy Notice</h2>
+      <h2>Terms of Service</h2>
       <p>
-        This Privacy Notice explains how Guardian Street Mapper collects, uses, shares, and protects information obtained from users of our website [Your Website URL] and any related services, applications, or features offered by us.
+        These Terms of Service ("Terms") govern your use of Guardian Street Mapper crime reporting website [Your Website URL] and any related services, applications, or features offered by us. By accessing or using the Website or Services, you agree to be bound by these Terms.
       </p>
-      <h3>Information We Collect</h3>
+      <h3>1. Acceptance of Terms</h3>
       <p>
-        <strong>Location Information:</strong> When you interact with our Website and use our Services to input the location of a crime incident onto a map, we may collect location information such as GPS coordinates or address details.
+        By accessing or using the Website or Services, you agree to be bound by these Terms. If you do not agree to these Terms, you may not access or use the Website or Services.
       </p>
+      <h3>2. Use of Services</h3>
       <p>
-        <strong>Crime Incident Details:</strong> In addition to location information, we may collect details of the crime incident that you input, including the type of crime, date and time of occurrence, description of the incident, and any other relevant information provided by you.
-      </p>
-      <h3>How We Use Your Information</h3>
-      <p>
-        We may use the information we collect from you for the following purposes:
+        You may use the Services only for lawful purposes and in accordance with these Terms. You agree not to:
       </p>
       <ul>
-        <li>Providing and improving our Services</li>
-        <li>Displaying crime incident data on the map for public awareness and safety purposes</li>
-        <li>Analyzing crime trends and patterns to enhance crime prevention efforts</li>
-        <li>Communicating with you regarding updates, alerts, or important information related to crime incidents or our Services</li>
+        <li>Violate any applicable laws, regulations, or third-party rights</li>
+        <li>Use the Services in any manner that could interfere with, disrupt, negatively affect, or inhibit other users from fully enjoying the Services</li>
+        <li>Attempt to gain unauthorized access to any part of the Services or bypass any measures we may use to prevent or restrict access to the Services</li>
       </ul>
-      <h3>Information Sharing</h3>
+      <h3>3. User Contributions</h3>
       <p>
-        We may share your information in the following circumstances:
+        You may have the opportunity to submit content or information through the Services, including crime reports or incident details. By submitting such content or information, you grant us a non-exclusive, perpetual, irrevocable, worldwide, royalty-free license to use, reproduce, modify, adapt, publish, translate, create derivative works from, distribute, and display such content or information in connection with the Services.
       </p>
-      <ul>
-        <li>With law enforcement agencies or authorities as required by law or to assist in investigations related to reported crime incidents</li>
-        <li>With third-party service providers who assist us in operating our Website and providing our Services</li>
-        <li>With your consent or as otherwise permitted or required by applicable law</li>
-      </ul>
-      <h3>Data Security</h3>
+      <h3>4. Intellectual Property Rights</h3>
       <p>
-        We take appropriate measures to protect the information we collect from unauthorized access, disclosure, alteration, or destruction. However, please be aware that no security measures are completely foolproof, and we cannot guarantee the security of your information.
+        The Website, Services, and their entire contents, features, and functionality (including but not limited to all information, software, text, displays, images, video, and audio) are owned by us or our licensors and are protected by copyright, trademark, patent, trade secret, and other intellectual property or proprietary rights laws.
       </p>
-      <h3>Your Choices</h3>
+      <h3>5. Limitation of Liability</h3>
       <p>
-        You have the option to voluntarily input information about crime incidents onto our Website. However, please consider the sensitivity of the information you provide and ensure that it does not include personal or identifying information of individuals involved in the crime incident, unless necessary and with their consent.
+        In no event shall we, our affiliates, directors, employees, agents, or licensors be liable for any indirect, incidental, special, consequential, or punitive damages, including but not limited to loss of profits, data, use, or other intangible losses, arising out of or relating to your use of or inability to use the Services.
       </p>
-      <h3>Children's Privacy</h3>
+      <h3>6. Governing Law</h3>
       <p>
-        Our Website and Services are not intended for children under the age of 13. We do not knowingly collect personal information from children under 13. If you are a parent or guardian and believe that your child has provided us with personal information, please contact us to request deletion.
+        These Terms shall be governed by and construed in accordance with the laws of Greater London, without regard to its conflict of law provisions.
       </p>
-      <h3>Changes to this Privacy Notice</h3>
+      <h3>7. Changes to Terms</h3>
       <p>
-        We reserve the right to update or change this Privacy Notice at any time. Any changes will be effective immediately upon posting the updated Privacy Notice on our Website.
+        We reserve the right to update or change these Terms at any time. Any changes will be effective immediately upon posting the updated Terms on the Website. Your continued use of the Website or Services after any such changes constitutes your acceptance of the new Terms.
       </p>
-      <h3>Contact Us</h3>
+      <h3>8. Contact Us</h3>
       <p>
-        If you have any questions or concerns about this Privacy Notice or our privacy practices, please contact us at [Your Contact Information].
+        If you have any questions or concerns about these Terms, please contact us at [Your Contact Information].
       </p>
     </div>
   );
 }
 
-export default PrivacyNotice;
+export default TermsOfService;
